v3.1 features updates

+++ Compressonator/Applications/CompressonatorGUI/Common/cvmatandqimage.cpp
@@ -403,7 +403,7 @@ QImage mat2Image_shared(const cv::Mat &mat, QImage::Format formatHint)
         }
     }
 
-    QImage img(mat.data, mat.cols, mat.rows, mat.step, formatHint);
+    QImage img(mat.data, mat.cols, mat.rows, (int)mat.step, formatHint);
 
     //Should we add directly support for user-customed-colorTable?
     if (formatHint == QImage::Format_Indexed8) {
